model/requirements.txt
absl-py==1.3.0
astunparse==1.6.3
attrs==22.1.0
autopep8==2.0.1
azure-cognitiveservices-speech==1.24.2
backcall==0.2.0
beautifulsoup4==4.11.1
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
cycler==0.11.0
decorator==4.4.2
flatbuffers==2.0.7
fonttools==4.38.0
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
importlib-metadata==5.1.0
ipython==7.34.0
jedi==0.18.2
joblib==1.2.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
media==0.1.6
mediapipe==0.9.0.1
mediapy==1.1.2
moviepy==1.0.3
mutagen==1.46.0
natsort==8.2.0
numpy==1.21.6
oauthlib==3.2.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==20.9
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
proglog==0.1.10
prompt-toolkit==3.0.36
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycryptodomex==3.16.0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scikit-multilearn==0.2.0
scipy==1.7.3
six==1.16.0
soupsieve==2.3.2.post1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-addons==0.19.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.28.0
tensorflowjs==3.19.0
termcolor==2.1.1
threadpoolctl==3.1.0
tomli==2.0.1
tqdm==4.64.1
traitlets==5.7.1
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.13
watchdog==2.2.0
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.2
wrapt==1.14.1
youtube-dl==2021.12.17
yt-dlp==2022.11.11
zipp==3.11.0
